fix: skip image-minimizer-webpack-plugin for RSPack builds
The image-minimizer-webpack-plugin is not compatible with RSPack's
lite-tapable implementation. It tries to tap into webpack-specific
hooks that don't exist in RSPack, causing builds to fail with:

  TypeError: Cannot read properties of undefined (reading 'tap')

Move ImageMinimizerPlugin configuration into bundler.js as a new
getImageMinimizers() function that conditionally returns:
- RSPack: empty array (images pass through unoptimized)
- webpack: the two ImageMinimizerPlugin instances (images + SVGs)
